+++ crates/wdl-lint/src/rules/doc_comment_tabs.rs
@@ -6,6 +6,7 @@ use wdl_analysis::LabeledSnippet;
 use wdl_analysis::Visitor;
 use wdl_ast::AstToken;
 use wdl_ast::Comment;
+use wdl_ast::CommentKind;
 use wdl_ast::Diagnostic;
 use wdl_ast::Span;
 use wdl_ast::SyntaxElement;
@@ -99,7 +100,7 @@ impl Visitor for DocCommentTabsRule {
     }
 
     fn comment(&mut self, diagnostics: &mut Diagnostics, comment: &Comment) {
-        if !comment.is_doc_comment() {
+        if comment.kind() != CommentKind::Documentation {
             return;
         }
         let text = comment.text();