+++ crates/wdl-lsp/CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.19.1 - 2026-05-14
 
+### Added
+
+* Added support for the `textDocument/foldingRange` request ([#867](https://github.com/stjude-rust-labs/sprocket/pull/867)).
+
 ## 0.19.0 - 2026-04-22
 
 ## 0.18.0 - 2026-04-02